move this to another process so we can see the other tests running
