+++ test/js/node/http/node-http.test.ts
@@ -1830,68 +1830,16 @@ if (process.platform !== "win32") {
   });
 }
 
-it("#10177 response.write with non-ascii latin1 should not cause duplicated character or segfault", done => {
-  // x = ascii
-  // á = latin1 supplementary character
-  // 📙 = emoji
-  // 👍🏽 = its a grapheme of 👍 🟤
-  // "\u{1F600}" = utf16
-  const chars = ["x", "á", "📙", "👍🏽", "\u{1F600}"];
-
-  // 128 = small than waterMark, 256 = waterMark, 1024 = large than waterMark
-  // 8Kb = small than cork buffer
-  // 16Kb = cork buffer
-  // 32Kb = large than cork buffer
-  const start_size = 128;
-  const increment_step = 1024;
-  const end_size = 32 * 1024;
-  let expected = "";
-
-  function finish(err) {
-    server.closeAllConnections();
-    Bun.gc(true);
-    done(err);
-  }
-  const server = require("http")
-    .createServer((_, response) => {
-      response.write(expected);
-      response.write("");
-      response.end();
-    })
-    .listen(0, "localhost", async (err, hostname, port) => {
-      expect(err).toBeFalsy();
-      expect(port).toBeGreaterThan(0);
-
-      for (const char of chars) {
-        for (let size = start_size; size <= end_size; size += increment_step) {
-          expected = char + Buffer.alloc(size, "-").toString("utf8") + "x";
-
-          try {
-            const url = `http://${hostname}:${port}`;
-            const count = 20;
-            const all = [];
-            const batchSize = 20;
-            while (all.length < count) {
-              const batch = Array.from({ length: batchSize }, () => fetch(url).then(a => a.text()));
-
-              all.push(...(await Promise.all(batch)));
-            }
-
-            using _ = disableAggressiveGCScope();
-            for (const result of all) {
-              expect(result).toBe(expected);
-            }
-          } catch (err) {
-            return finish(err);
-          }
-        }
-
-        // still always run GC at the end here.
-        Bun.gc(true);
-      }
-      finish();
-    });
-}, 20_000);
+it("#10177 response.write with non-ascii latin1 should not cause duplicated character or segfault", () => {
+  // this can cause a segfault so we run it in a separate process
+  const { exitCode } = Bun.spawnSync({
+    cmd: [bunExe(), "run", path.join(import.meta.dir, "node-http-response-write-encode-fixture.js")],
+    env: bunEnv,
+    stdout: "inherit",
+    stderr: "inherit",
+  });
+  expect(exitCode).toBe(0);
+}, 60_000);
 
 it("#11425 http no payload limit", done => {
   const server = Server((req, res) => {
